pyproject.toml
[tool.poetry]
name = "discord-chan"
version = "2.0.0"
description = "General purpose Discord bot"
authors = ["StarrFox <starrfox6312@gmail.com>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.0.1"
jishaku = "^2.5.1"
numpy = "^1.23.2"
aiosqlite = "^0.17.0"
uwuify = "^1.1.0"
pyenchant = "^3.2.2"
click = "^8.1.3"
click-default-group = "^1.2.2"
loguru = "^0.6.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
humanize = "^4.3.0"
python-dotenv = "^0.20.0"
psutil = "^5.9.1"
sqlmodel = "^0.0.8"
appdirs = "^1.4.4"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
discord_chan = "discord_chan.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
